pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "utilityscripts"
authors = [{name = "Sean Kane", email = "mail@ifitsmoving.com"}]
description = "Various utility scripts"
readme = "README.md"
license = {file = "LICENSE.md"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version"]
requires-python = ">=3.12"
dependencies = [
 "numpy",
 "matplotlib",
 "pint",
 "polars",
 "ipython",
 "numpy",
 "sympy",
 "pint",
 "shapely",
 "concreteproperties",
 "sectionproperties",
 "simplebeam",
 "Humre",
 "scipy",
 "toml",
 "PyniteFEA",
 "altair",
 "openpyxl",
 "fastexcel",
 "Pillow",
 "pillow-heif",
 "pandas",
 "python-docx",
]

[project.urls]
Home = "https://ifitsmoving.com"

[project.optional-dependencies]
dev = [
 'isort',
 'coverage',
 'mypy',
 'pylint',
 'pytest',
 'ruff',
 "pytest-randomly",
 "pytest-cov",
 "pytest-sugar",
 "pre-commit",
 "types-shapely",
 "pandas-stubs",
 "types-tqdm",
]

[tool.uv]
package = true

[tool.hatch.version]
path = "src/utilityscripts/__init__.py"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
utilityscripts = ["**.xlsx"]

[tool.pylint]
disable = ["C0103"]
